refactor(WebAPI): Moved MoveFileWithResumeCommand to Application project
